% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p5.R
\name{p5}
\alias{p5}
\title{Create a \code{p5.js} htmlwidget.}
\usage{
p5(code, width = NULL, height = NULL)
}
\arguments{
\item{code}{Either a file with JavaScript code or a string with JavaScript}

\item{width}{width}

\item{height}{height}
}
\value{

}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Create a \code{p5.js} htmlwidget
}
